+++ CHANGELOG.md
@@ -18,3 +18,9 @@ Change Log
 03.23.2019
 ----------
 * Updated core libraries
+
+03.30.2019
+----------
+* Updated core libraries
+* Refactored to support new core libraries
+* Updated docker build to account for new directory structure